+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Enhancements
 
 - Do not start new process manager instance on `:continue` ([#181](https://github.com/commanded/commanded/pull/181)).
+- Support [`Phoenix.PubSub` v1.1.0](https://hexdocs.pm/phoenix/1.1.0/Phoenix.PubSub.html).
+
 
 ## v0.17.0